Move the auth thunk files into the auth directory, according to the role
