+++ frontend/src/__tests__/store/auth/thunks/deleteAccount.test.ts
@@ -5,7 +5,7 @@ import {
   fetchAuthUser,
   SignInRequest,
   signInWithEmail,
-} from 'store/thunks';
+} from 'store/thunks/auth';
 import { initializeStore, store } from 'mocks/utils/store';
 import {
   getFlashState,
