+++ frontend/src/__tests__/store/auth/thunks/signOutFromAPI.test.ts
@@ -5,7 +5,7 @@ import {
   SignInRequest,
   signInWithEmail,
   signOutFromAPI,
-} from 'store/thunks';
+} from 'store/thunks/auth';
 import { initializeStore, store } from 'mocks/utils/store';
 import {
   getFlashState,
