+++ frontend/src/__tests__/store/auth/thunks/updatePassword.test.ts
@@ -5,7 +5,7 @@ import {
   signInWithEmail,
   updatePassword,
   UpdatePasswordRequest,
-} from 'store/thunks';
+} from 'store/thunks/auth';
 import { initializeStore, store } from 'mocks/utils/store';
 import {
   getFlashState,
