+++ frontend/src/mocks/controllers/auth/updateProfile.ts
@@ -1,4 +1,4 @@
-import { UpdateProfileRequest, UpdateProfileResponse } from 'store/thunks';
+import { UpdateProfileRequest, UpdateProfileResponse } from 'store/thunks/auth';
 import { db, UserDocument } from 'mocks/models';
 
 type UpdateProfileProps = {
