+++ frontend/src/mocks/handlers.ts
@@ -12,7 +12,7 @@ import {
   UpdatePasswordRequest,
   ForgotPasswordRequest,
   ResetPasswordRequest,
-} from 'store/thunks';
+} from 'store/thunks/auth';
 import { auth, db, sanitizeUser } from 'mocks/models';
 import {
   createUserController,
