+++ frontend/src/store/thunks/auth/deleteAccount.ts
