+++ frontend/src/store/thunks/auth/fetchAuthUser.ts
