+++ frontend/src/store/thunks/auth/signOutFromAPI.ts
